Fix Zip-Slip, AppleDouble overflow, and .DS_Store double-free (CR-1,2,3,9)
Security & crash must-fix from the PR #100 review:

- GWMetaArchive extractArchive:toDir: rejected no entry names, so a
  crafted zip with "../" or absolute paths could write outside the
  destination (Zip Slip). Add path_is_within() and skip any entry whose
  standardized destPath escapes the canonical destDir. Also skip entries
  whose name is not valid UTF-8 (epath == nil) before building destPath.

- GSAppleDouble initWithData: computed dataOff + dataLen in 32-bit
  unsigned arithmetic, so a crafted descriptor could wrap the sum and
  slip past the bounds check into an out-of-bounds read. Widen the
  addition to 64-bit.

- GWViewSettingsManager readSettings released an autoreleased DSStoreInfo
  (+infoForDirectoryPath: is autoreleased) when a folder's .DS_Store
  existed but failed to load -> double-free on pool drain. Drop the
  reference without releasing.

+++ GWMetadata/GSAppleDouble.m
@@ -132,8 +132,11 @@ writeBE16(uint8_t *bytes, uint16_t value)
       uint32_t dataOff  = readBE32(desc + 4);
       uint32_t dataLen  = readBE32(desc + 8);
 
-      /* Validate offset and length */
-      if (dataOff + dataLen > length || dataOff < descriptorsEnd)
+      /* Validate offset and length.  Compute the sum in 64-bit so a
+       * crafted (dataOff, dataLen) cannot wrap the 32-bit addition and
+       * slip past this bound check into an out-of-bounds read below. */
+      if ((uint64_t)dataOff + (uint64_t)dataLen > length
+          || dataOff < descriptorsEnd)
         {
           /* Malformed entry; skip it rather than failing entirely */
           continue;
